replace new_notification with notification
cleaned up code
